# backend/app/src/gen/pb/api/v2/app/user/service/UserService_rb.controller.ts
// @generated by protoc-gen-restobuf-nestjs v1 with parameter "target=ts,import_extension=js"
// @generated from file api/v2/app/user/service.proto (package api.v2.app.user, syntax proto3)
/* eslint-disable */

import { Body, Controller, Get, Param, Post, Query, Req, Res } from "@nestjs/common";
import { UserServiceService } from "./UserService_rb.service";
import type { JsonObject, JsonValue } from "@bufbuild/protobuf";
import { fromJson, toJson } from "@bufbuild/protobuf";
import { Request, Response } from "express";
import { CreateUserRequestSchema, CreateUserResponseSchema, GetUserRequestSchema, GetUserResponseSchema } from "../service_pb.js";

@Controller() 
export class UserServiceController {
  constructor(private service: UserServiceService) {}

  @Post('/api/v2/app/user')
  async handleCreateUser(
    @Param() pathParams: {[key: string]: string},
    @Query() queryParams: {[key: string]: string},
    @Body() body: JsonObject,
    @Req() req: Request,
    @Res({ passthrough: true}) res: Response,
  ): Promise<JsonValue> {
      body ??= {};
    for (const key in queryParams) {
      setMessageField(body, key.split('.'), queryParams[key]);
    }
    const input = fromJson(CreateUserRequestSchema, body);
    const output = await this.service.handleCreateUser(input, req, res);
    return toJson(CreateUserResponseSchema, output);
  }

  @Get('/api/v2/app/user/:user_id')
  async handleGetUser(
    @Param() pathParams: {[key: string]: string},
    @Query() queryParams: {[key: string]: string},
    @Body() body: JsonObject,
    @Req() req: Request,
    @Res({ passthrough: true}) res: Response,
  ): Promise<JsonValue> {
      body ??= {};
    for (const key in queryParams) {
      setMessageField(body, key.split('.'), queryParams[key]);
    }
    setMessageField(body, ["user_id"], pathParams['user_id']);
    const input = fromJson(GetUserRequestSchema, body);
    const output = await this.service.handleGetUser(input, req, res);
    return toJson(GetUserResponseSchema, output);
  }

}

function setMessageField(message: JsonObject, fieldPath: string[], value: string): JsonValue {
    if(fieldPath.length === 0) {
        throw new Error("fieldPath must not be empty");
    }
    const [field, ...rest] = fieldPath;
    if (fieldPath.length === 1) {
        message[field] = value;
        return message;
    }else {
        message[field] ??= {};
        return setMessageField(message[field] as JsonObject , rest, value);
    }
}