# backend/app/src/gen/pb/api/v2/app/user/service_pb.ts
// @generated by protoc-gen-es v2.2.2 with parameter "target=ts,import_extension=js,json_types=true"
// @generated from file api/v2/app/user/service.proto (package api.v2.app.user, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage, GenService } from "@bufbuild/protobuf/codegenv1";
import { fileDesc, messageDesc, serviceDesc } from "@bufbuild/protobuf/codegenv1";
import { file_google_api_annotations } from "../../../../google/api/annotations_pb.js";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file api/v2/app/user/service.proto.
 */
export const file_api_v2_app_user_service: GenFile = /*@__PURE__*/
  fileDesc("Ch1hcGkvdjIvYXBwL3VzZXIvc2VydmljZS5wcm90bxIPYXBpLnYyLmFwcC51c2VyIikKEUNyZWF0ZVVzZXJSZXF1ZXN0EhQKDGRpc3BsYXlfbmFtZRgBIAEoCSJRChJDcmVhdGVVc2VyUmVzcG9uc2USDwoHdXNlcl9pZBgBIAEoCRIUCgxkaXNwbGF5X25hbWUYAiABKAkSFAoMc2VjcmV0X3Rva2VuGAMgASgJIiEKDkdldFVzZXJSZXF1ZXN0Eg8KB3VzZXJfaWQYASABKAkiOAoPR2V0VXNlclJlc3BvbnNlEg8KB3VzZXJfaWQYASABKAkSFAoMZGlzcGxheV9uYW1lGAIgASgJMvABCgtVc2VyU2VydmljZRJvCgpDcmVhdGVVc2VyEiIuYXBpLnYyLmFwcC51c2VyLkNyZWF0ZVVzZXJSZXF1ZXN0GiMuYXBpLnYyLmFwcC51c2VyLkNyZWF0ZVVzZXJSZXNwb25zZSIYgtPkkwISIhAvYXBpL3YyL2FwcC91c2VyEnAKB0dldFVzZXISHy5hcGkudjIuYXBwLnVzZXIuR2V0VXNlclJlcXVlc3QaIC5hcGkudjIuYXBwLnVzZXIuR2V0VXNlclJlc3BvbnNlIiKC0+STAhwSGi9hcGkvdjIvYXBwL3VzZXIve3VzZXJfaWR9QoYBChNjb20uYXBpLnYyLmFwcC51c2VyQgxTZXJ2aWNlUHJvdG9QAVoBLqICBEFWQVWqAg9BcGkuVjIuQXBwLlVzZXLKAg9BcGlcVjJcQXBwXFVzZXLiAhtBcGlcVjJcQXBwXFVzZXJcR1BCTWV0YWRhdGHqAhJBcGk6OlYyOjpBcHA6OlVzZXJiBnByb3RvMw", [file_google_api_annotations]);

/**
 * @generated from message api.v2.app.user.CreateUserRequest
 */
export type CreateUserRequest = Message<"api.v2.app.user.CreateUserRequest"> & {
  /**
   * @generated from field: string display_name = 1;
   */
  displayName: string;
};

/**
 * @generated from message api.v2.app.user.CreateUserRequest
 */
export type CreateUserRequestJson = {
  /**
   * @generated from field: string display_name = 1;
   */
  displayName?: string;
};

/**
 * Describes the message api.v2.app.user.CreateUserRequest.
 * Use `create(CreateUserRequestSchema)` to create a new message.
 */
export const CreateUserRequestSchema: GenMessage<CreateUserRequest, CreateUserRequestJson> = /*@__PURE__*/
  messageDesc(file_api_v2_app_user_service, 0);

/**
 * @generated from message api.v2.app.user.CreateUserResponse
 */
export type CreateUserResponse = Message<"api.v2.app.user.CreateUserResponse"> & {
  /**
   * @generated from field: string user_id = 1;
   */
  userId: string;

  /**
   * @generated from field: string display_name = 2;
   */
  displayName: string;

  /**
   * @generated from field: string secret_token = 3;
   */
  secretToken: string;
};

/**
 * @generated from message api.v2.app.user.CreateUserResponse
 */
export type CreateUserResponseJson = {
  /**
   * @generated from field: string user_id = 1;
   */
  userId?: string;

  /**
   * @generated from field: string display_name = 2;
   */
  displayName?: string;

  /**
   * @generated from field: string secret_token = 3;
   */
  secretToken?: string;
};

/**
 * Describes the message api.v2.app.user.CreateUserResponse.
 * Use `create(CreateUserResponseSchema)` to create a new message.
 */
export const CreateUserResponseSchema: GenMessage<CreateUserResponse, CreateUserResponseJson> = /*@__PURE__*/
  messageDesc(file_api_v2_app_user_service, 1);

/**
 * @generated from message api.v2.app.user.GetUserRequest
 */
export type GetUserRequest = Message<"api.v2.app.user.GetUserRequest"> & {
  /**
   * @generated from field: string user_id = 1;
   */
  userId: string;
};

/**
 * @generated from message api.v2.app.user.GetUserRequest
 */
export type GetUserRequestJson = {
  /**
   * @generated from field: string user_id = 1;
   */
  userId?: string;
};

/**
 * Describes the message api.v2.app.user.GetUserRequest.
 * Use `create(GetUserRequestSchema)` to create a new message.
 */
export const GetUserRequestSchema: GenMessage<GetUserRequest, GetUserRequestJson> = /*@__PURE__*/
  messageDesc(file_api_v2_app_user_service, 2);

/**
 * @generated from message api.v2.app.user.GetUserResponse
 */
export type GetUserResponse = Message<"api.v2.app.user.GetUserResponse"> & {
  /**
   * @generated from field: string user_id = 1;
   */
  userId: string;

  /**
   * @generated from field: string display_name = 2;
   */
  displayName: string;
};

/**
 * @generated from message api.v2.app.user.GetUserResponse
 */
export type GetUserResponseJson = {
  /**
   * @generated from field: string user_id = 1;
   */
  userId?: string;

  /**
   * @generated from field: string display_name = 2;
   */
  displayName?: string;
};

/**
 * Describes the message api.v2.app.user.GetUserResponse.
 * Use `create(GetUserResponseSchema)` to create a new message.
 */
export const GetUserResponseSchema: GenMessage<GetUserResponse, GetUserResponseJson> = /*@__PURE__*/
  messageDesc(file_api_v2_app_user_service, 3);

/**
 * @generated from service api.v2.app.user.UserService
 */
export const UserService: GenService<{
  /**
   * @generated from rpc api.v2.app.user.UserService.CreateUser
   */
  createUser: {
    methodKind: "unary";
    input: typeof CreateUserRequestSchema;
    output: typeof CreateUserResponseSchema;
  },
  /**
   * @generated from rpc api.v2.app.user.UserService.GetUser
   */
  getUser: {
    methodKind: "unary";
    input: typeof GetUserRequestSchema;
    output: typeof GetUserResponseSchema;
  },
}> = /*@__PURE__*/
  serviceDesc(file_api_v2_app_user_service, 0);